requirements.txt
Flask==3.0.0
Pillow==10.1.0
supabase==2.4.6
python-dotenv==1.0.0
